Move timeout handler execution to ComponentTimeoutExecutor
